# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OM PAY API",
        "description": "API de paiement mobile OM PAY avec authentification OTP et gestion de comptes",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement"
        },
        {
            "url": "https://ompay-isuf.onrender.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/compte/{compte}/solde": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Consulter le solde du compte",
                "description": "Retourne le solde actuel du compte spécifié",
                "operationId": "getSolde",
                "parameters": [
                    {
                        "name": "compte",
                        "in": "path",
                        "description": "ID du compte (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "a054a43b-8206-4efe-914b-c1f96457d58e"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solde du compte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "solde": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 15000.5
                                                },
                                                "numero_compte": {
                                                    "type": "string",
                                                    "example": "OM-2025-AB12-CD34"
                                                },
                                                "devise": {
                                                    "type": "string",
                                                    "example": "FCFA"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès non autorisé à ce compte"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/compte/{compte}/depot": {
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Effectuer un dépôt d'argent",
                "description": "Crédite le compte spécifié avec le montant indiqué",
                "operationId": "depot",
                "parameters": [
                    {
                        "name": "compte",
                        "in": "path",
                        "description": "ID du compte (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "a054a43b-8206-4efe-914b-c1f96457d58e"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "montant"
                                ],
                                "properties": {
                                    "montant": {
                                        "description": "Montant à déposer (min: 100, max: 1 000 000)",
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dépôt effectué avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dépôt effectué avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "transaction": {
                                                    "properties": {
                                                        "reference": {
                                                            "type": "string",
                                                            "example": "TXN-20251110-ABC123"
                                                        },
                                                        "type": {
                                                            "type": "string",
                                                            "example": "depot"
                                                        },
                                                        "montant": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 5000
                                                        },
                                                        "statut": {
                                                            "type": "string",
                                                            "example": "reussi"
                                                        },
                                                        "compte_destinataire": {
                                                            "properties": {
                                                                "numero_compte": {
                                                                    "type": "string",
                                                                    "example": "OM-2025-AB12-CD34"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "nouveau_solde": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 20000.5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Montant invalide ou erreur de traitement"
                    },
                    "403": {
                        "description": "Accès non autorisé à ce compte"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/compte/{compte}/qrcode": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Récupérer l'image QR code du compte",
                "description": "Retourne l'URL de l'image QR code du compte pour faciliter les paiements",
                "operationId": "getQrCode",
                "parameters": [
                    {
                        "name": "compte",
                        "in": "path",
                        "description": "ID du compte (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "a054a43b-8206-4efe-914b-c1f96457d58e"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "URL du QR code",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "qr_code_url": {
                                                    "type": "string",
                                                    "format": "uri",
                                                    "example": "https://ompay.onrender.com/storage/qrcodes/qr_OM-2025-AB12-CD34_abc123.png"
                                                },
                                                "numero_compte": {
                                                    "type": "string",
                                                    "example": "OM-2025-AB12-CD34"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès non autorisé à ce compte"
                    },
                    "404": {
                        "description": "QR code non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}